+++ src/locale-config.ts
