+++ src/navigation.ts
@@ -1,5 +1,5 @@
 import { createLocalizedPathnamesNavigation } from "next-intl/navigation";
-import { localePrefix, locales, pathnames } from "./config";
+import { localePrefix, locales, pathnames } from "./locale-config";
 
 export const { Link, redirect, usePathname, useRouter } =
   createLocalizedPathnamesNavigation({
